Implement LocalJtxCollectionStore
