+++ app/src/main/kotlin/at/bitfire/davdroid/resource/LocalAddressBookStore.kt
@@ -20,10 +20,10 @@ import at.bitfire.davdroid.repository.DavCollectionRepository
 import at.bitfire.davdroid.repository.DavServiceRepository
 import at.bitfire.davdroid.resource.LocalAddressBook.Companion.USER_DATA_COLLECTION_ID
 import at.bitfire.davdroid.resource.LocalAddressBook.Companion.USER_DATA_URL
-import at.bitfire.davdroid.resource.LocalAddressBook.Companion.accountName
 import at.bitfire.davdroid.settings.Settings
 import at.bitfire.davdroid.settings.SettingsManager
 import at.bitfire.davdroid.sync.account.SystemAccountUtils
+import at.bitfire.davdroid.util.DavUtils.lastSegment
 import at.bitfire.davdroid.util.setAndVerifyUserData
 import dagger.hilt.android.qualifiers.ApplicationContext
 import java.util.logging.Level
@@ -35,18 +35,48 @@ class LocalAddressBookStore @Inject constructor(
     val addressBookFactory: LocalAddressBook.Factory,
     val collectionRepository: DavCollectionRepository,
     @ApplicationContext val context: Context,
+    val localAddressBookFactory: LocalAddressBook.Factory,
     val logger: Logger,
-    val settings: SettingsManager,
-    val serviceRepository: DavServiceRepository
-): LocalDataStore<LocalAddressBook> {
+    val serviceRepository: DavServiceRepository,
+    val settings: SettingsManager
+    ): LocalDataStore<LocalAddressBook> {
 
     /** whether a (usually managed) setting wants all address-books to be read-only **/
     val forceAllReadOnly: Boolean
         get() = settings.getBoolean(Settings.FORCE_READ_ONLY_ADDRESSBOOKS)
 
 
+    /**
+     * Assembles a name for the address book (account) from its corresponding database [Collection].
+     *
+     * The address book account name contains
+     *
+     * - the collection display name or last URL path segment
+     * - the actual account name
+     * - the collection ID, to make it unique.
+     *
+     * @param info  Collection to take info from
+     */
+    fun accountName(info: Collection): String {
+        // Name the address book after given collection display name, otherwise use last URL path segment
+        val sb = StringBuilder(info.displayName.let {
+            if (it.isNullOrEmpty())
+                info.url.lastSegment
+            else
+                it
+        })
+        // Add the actual account name to the address book account name
+        serviceRepository.get(info.serviceId)?.let { service ->
+            sb.append(" (${service.accountName})")
+        }
+        // Add the collection ID for uniqueness
+        sb.append(" #${info.id}")
+        return sb.toString()
+    }
+
+
     override fun create(provider: ContentProviderClient, fromCollection: Collection): LocalAddressBook? {
-        val name = accountName(context, fromCollection)
+        val name = accountName(fromCollection)
         val account = createAccount(
             name = name,
             id = fromCollection.id,
@@ -78,12 +108,41 @@ class LocalAddressBookStore @Inject constructor(
         return account
     }
 
+
+    override fun getAll(account: Account, provider: ContentProviderClient): List<LocalAddressBook> =
+        serviceRepository.getByAccountAndType(account.name, Service.TYPE_CARDDAV)?.let { service ->
+            // get all collections for the CardDAV service
+            collectionRepository.getByService(service.id).mapNotNull { collection ->
+                // and map to a LocalAddressBook, if applicable
+                findByCollection(provider, collection.id)
+            }
+        }.orEmpty()
+
+    /**
+     * Finds a [LocalAddressBook] based on its corresponding collection.
+     *
+     * @param id    collection ID to look for
+     *
+     * @return The [LocalAddressBook] for the given collection or *null* if not found
+     */
+    private fun findByCollection(provider: ContentProviderClient, id: Long): LocalAddressBook? {
+        val accountManager = AccountManager.get(context)
+        return accountManager
+            .getAccountsByType(context.getString(R.string.account_type_address_book))
+            .filter { account ->
+                accountManager.getUserData(account, USER_DATA_COLLECTION_ID)?.toLongOrNull() == id
+            }
+            .map { account -> localAddressBookFactory.create(account, provider) }
+            .firstOrNull()
+    }
+
+
     override fun update(provider: ContentProviderClient, localCollection: LocalAddressBook, fromCollection: Collection) {
         var currentAccount = localCollection.addressBookAccount
         logger.log(Level.INFO, "Updating local address book $currentAccount from collection $fromCollection")
 
         // Update the account name
-        val newAccountName = accountName(context, fromCollection)
+        val newAccountName = accountName(fromCollection)
         if (currentAccount.name != newAccountName) {
             // rename, move contacts/groups and update [AndroidAddressBook.]account
             localCollection.renameAccount(newAccountName)
@@ -130,16 +189,24 @@ class LocalAddressBookStore @Inject constructor(
         localCollection.updateSyncFrameworkSettings()
     }
 
-    override fun getAll(account: Account, provider: ContentProviderClient): List<LocalAddressBook> =
-        serviceRepository.getByAccountAndType(account.name, Service.TYPE_CARDDAV)?.let { service ->
-            collectionRepository.getByService(service.id).mapNotNull { collection ->
-                LocalAddressBook.findByCollection(context, provider, collection.id)
-            }
-        }.orEmpty()
 
     override fun delete(localCollection: LocalAddressBook) {
         val accountManager = AccountManager.get(context)
         accountManager.removeAccountExplicitly(localCollection.addressBookAccount)
+    }
+
+    /**
+     * Deletes a [LocalAddressBook] based on its corresponding database collection.
+     *
+     * @param id    [Collection.id] to look for
+     */
+    fun deleteByCollectionId(id: Long) {
+        val accountManager = AccountManager.get(context)
+        val addressBookAccount = accountManager.getAccountsByType(context.getString(R.string.account_type_address_book)).firstOrNull { account ->
+            accountManager.getUserData(account, USER_DATA_COLLECTION_ID)?.toLongOrNull() == id
+        }
+        if (addressBookAccount != null)
+            accountManager.removeAccountExplicitly(addressBookAccount)
     }
