+++ app/src/main/kotlin/at/bitfire/davdroid/sync/JtxSyncer.kt
@@ -10,15 +10,12 @@ import android.content.ContentProviderClient
 import android.os.Build
 import at.bitfire.davdroid.db.Collection
 import at.bitfire.davdroid.db.Service
-import at.bitfire.davdroid.repository.PrincipalRepository
 import at.bitfire.davdroid.resource.LocalJtxCollection
 import at.bitfire.davdroid.resource.LocalJtxCollectionStore
-import at.bitfire.ical4android.JtxCollection
 import at.bitfire.ical4android.TaskProvider
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
-import java.util.logging.Level
 
 /**
  * Sync logic for jtx board
@@ -27,8 +24,8 @@ class JtxSyncer @AssistedInject constructor(
     @Assisted account: Account,
     @Assisted extras: Array<String>,
     @Assisted syncResult: SyncResult,
+    localJtxCollectionStore: LocalJtxCollectionStore,
     private val jtxSyncManagerFactory: JtxSyncManager.Factory,
-    private val principalRepository: PrincipalRepository,
     private val tasksAppManager: dagger.Lazy<TasksAppManager>
 ): Syncer<LocalJtxCollectionStore, LocalJtxCollection>(account, extras, syncResult) {
 
@@ -37,17 +34,13 @@ class JtxSyncer @AssistedInject constructor(
         fun create(account: Account, extras: Array<String>, syncResult: SyncResult): JtxSyncer
     }
 
-    override val dataStore: LocalJtxCollectionStore
-        get() = TODO("Not yet implemented")
+    override val dataStore = localJtxCollectionStore
 
     override val serviceType: String
         get() = Service.TYPE_CALDAV
     override val authority: String
         get() = TaskProvider.ProviderName.JtxBoard.authority
 
-
-    /*override fun getLocalCollections(provider: ContentProviderClient): List<LocalJtxCollection>
-        = JtxCollection.find(account, provider, context, LocalJtxCollection.Factory, null, null)*/
 
     override fun prepare(provider: ContentProviderClient): Boolean {
         // check whether jtx Board is new enough
@@ -73,26 +66,6 @@ class JtxSyncer @AssistedInject constructor(
 
     override fun getDbSyncCollections(serviceId: Long): List<Collection> =
         collectionRepository.getSyncJtxCollections(serviceId)
-
-    /*override fun update(localCollection: LocalJtxCollection, remoteCollection: Collection) {
-        logger.log(Level.FINE, "Updating local jtx collection ${remoteCollection.url}", remoteCollection)
-        val owner = remoteCollection.ownerId?.let { principalRepository.get(it) }
-        localCollection.updateCollection(remoteCollection, owner, accountSettings.getManageCalendarColors())
-    }*/
-
-    /*override fun create(provider: ContentProviderClient, remoteCollection: Collection): LocalJtxCollection {
-        logger.log(Level.INFO, "Adding local jtx collection", remoteCollection)
-        val owner = remoteCollection.ownerId?.let { principalRepository.get(it) }
-        val uri = LocalJtxCollection.create(account, provider, remoteCollection, owner)
-        return JtxCollection.find(
-            account,
-            provider,
-            context,
-            LocalJtxCollection.Factory,
-            "${JtxContract.JtxCollection.ID} = ?",
-            arrayOf("${ContentUris.parseId(uri)}")
-        ).first()
-    }*/
 
     override fun syncCollection(provider: ContentProviderClient, localCollection: LocalJtxCollection, remoteCollection: Collection) {
         logger.info("Synchronizing jtx collection $localCollection")
